requirements.txt
asgiref==3.5.0
certifi==2021.10.8
charset-normalizer==2.0.12
coreapi==2.3.3
coreschema==0.0.4
Django==4.0.2
django-environ==0.8.1
django-phonenumber-field==6.0.0
django-phonenumbers==1.0.1
djangorestframework==3.12.4
djangorestframework-simplejwt==5.0.0
drf-yasg==1.20.0
idna==3.3
inflection==0.5.1
itypes==1.2.0
Jinja2==3.0.3
MarkupSafe==2.1.0
packaging==21.3
phonenumbers==8.12.44
Pillow==9.0.1
psycopg2-binary==2.9.3
PyJWT==2.3.0
pyparsing==3.0.7
requests==2.27.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
sqlparse==0.4.2
uritemplate==4.1.1
urllib3==1.26.8
